server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
motor==3.7.1
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pymongo==4.15.3
pyparsing==3.2.5
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
wheel==0.45.1
